requirements.txt
Django==2.2.10
djangorestframework==3.10.2
djangorestframework-stubs==0.4.2
six==1.11.0
requests==2.20.0
setuptools==54.2.0
urllib3==1.24.2
pytz==2021.1
sqlparse==0.4.1
django-bootstrap4==1.1.1
